chore: consolidate env and move helper functions to utils.go
